Add file position tracking
Adds a system to track global file positions in a set of files
and adds position infomration to the nodes. A file set should be
a singleton for the compilation invocation.

This is inspired by a similar system used by Go's language parsing
modules which is also simmilar to a system I built at Borland for
a private source C# parser used in C# Builder.