Simplified the the language to be more IR like
Removed higher level constructs as well as focus more on physcial
constructs. For example, removed all notion of `var` and `val` which
can be created as a transform into this lower level constructs. In
other words if a construct can be represented as a lower level construct
it should be removed from the language and implemented as a transform.

This is not complete yet as and the AST contains higher level nodes.
I have yet to decide if these should be removed or isolated.